Extract pocket-tts-mcp to standalone PyPI package
Split the monolithic mcp-pockettts into two servers:
- pocket-tts-mcp: published to PyPI, installed via uvx/pip
- mcp-recording: minimal stop_recording tool for demo container
